+++ packages/realm-web/src/UserProfile.ts
@@ -16,6 +16,16 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
+/**
+ * The type of a user.
+ */
+enum UserType {
+    /** A normal end-user created this user */
+    Normal = "normal",
+    /** The user was created by the server */
+    Server = "server",
+}
+
 /** @ignore */
 enum DataKey {
     /** @ignore */
@@ -80,63 +90,52 @@ export class UserProfile implements Realm.UserProfile {
     public readonly maxAge?: string;
 
     /** @inheritdoc */
-    public readonly type: Realm.UserType;
+    public readonly type: Realm.UserType = UserType.Normal;
 
     /** @inheritdoc */
-    public readonly identities: Realm.UserIdentity[];
+    public readonly identities: Realm.UserIdentity[] = [];
 
     /**
      * Construct a user profile from the body of a response
      *
      * @param response The response of a call fetching the users profile
      */
-    constructor(response: any) {
-        /**
-      -  "data": {}
-      -  "domain_id": "5ed10debc085000e2c0097ac"
-      -  "identities": [
-      -    {
-      -      "id": "5ed10e0dc085000e2c0099f2-fufttusvpmojykvacvhijoaq"
-      -      "provider_id": "5ed10dedc085000e2c0097c5"
-      -      "provider_type": "anon-user"
-      -    }
-      -  ]
-      -  "type": "normal"
-      -  "user_id": "5ed10e0dc085000e2c0099f3"
-         */
-        if (typeof response.type === "string") {
-            this.type = response.type;
-        } else {
-            throw new Error("Expected 'type' in the response body");
-        }
-
-        if (Array.isArray(response.identities)) {
-            this.identities = response.identities.map((identity: any) => {
-                return {
-                    id: identity.id,
-                    providerId: identity["provider_id"],
-                    providerType: identity["provider_type"],
-                };
-            });
-        } else {
-            throw new Error("Expected 'identities' in the response body");
-        }
-
-        const { data } = response;
-        if (typeof data === "object") {
-            for (const key in DATA_MAPPING) {
-                const value = data[key];
-                const propertyName = DATA_MAPPING[key as DataKey];
-                if (
-                    typeof value === "string" &&
-                    propertyName !== "identities" &&
-                    propertyName !== "type"
-                ) {
-                    this[propertyName] = value;
-                }
+    constructor(response?: any) {
+        if (response) {
+            if (typeof response.type === "string") {
+                this.type = response.type;
+            } else {
+                throw new Error("Expected 'type' in the response body");
             }
-        } else {
-            throw new Error("Expected 'data' in the response body");
+
+            if (Array.isArray(response.identities)) {
+                this.identities = response.identities.map((identity: any) => {
+                    return {
+                        id: identity.id,
+                        providerId: identity["provider_id"],
+                        providerType: identity["provider_type"],
+                    };
+                });
+            } else {
+                throw new Error("Expected 'identities' in the response body");
+            }
+
+            const { data } = response;
+            if (typeof data === "object") {
+                for (const key in DATA_MAPPING) {
+                    const value = data[key];
+                    const propertyName = DATA_MAPPING[key as DataKey];
+                    if (
+                        typeof value === "string" &&
+                        propertyName !== "identities" &&
+                        propertyName !== "type"
+                    ) {
+                        this[propertyName] = value;
+                    }
+                }
+            } else {
+                throw new Error("Expected 'data' in the response body");
+            }
         }
     }
 }